KAFKA-18704: Move ScramPublisher to metadata module

# Conflicts:
#	core/src/main/scala/kafka/server/BrokerServer.scala
#	metadata/src/main/java/org/apache/kafka/metadata/publisher/ScramPublisher.java
#	server-common/src/main/java/org/apache/kafka/security/CredentialProvider.java
#	server-common/src/main/java/org/apache/kafka/server/common/CredentialProvider.java
#	server/src/main/java/org/apache/kafka/security/CredentialProvider.java

# Conflicts:
#	core/src/main/scala/kafka/tools/TestRaftServer.scala
