+++ metadata/src/main/java/org/apache/kafka/metadata/publisher/ScramPublisher.java
@@ -16,15 +16,21 @@
  */
 package org.apache.kafka.metadata.publisher;
 
+import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.image.MetadataDelta;
 import org.apache.kafka.image.MetadataImage;
 import org.apache.kafka.image.ScramDelta;
 import org.apache.kafka.image.loader.LoaderManifest;
 import org.apache.kafka.image.publisher.MetadataPublisher;
-import org.apache.kafka.security.CredentialProvider;
+import org.apache.kafka.server.common.CredentialProvider;
 import org.apache.kafka.server.fault.FaultHandler;
 
+import org.slf4j.Logger;
+
+import java.util.Optional;
+
 public class ScramPublisher implements MetadataPublisher {
+    private final Logger log;
     private final int nodeId;
     private final FaultHandler faultHandler;
     private final String nodeType;
@@ -35,6 +41,7 @@ public class ScramPublisher implements MetadataPublisher {
         this.faultHandler = faultHandler;
         this.nodeType = nodeType;
         this.credentialProvider = credentialProvider;
+        this.log = new LogContext(name()).logger(ScramPublisher.class);
     }
 
     @Override
@@ -54,6 +61,7 @@ public class ScramPublisher implements MetadataPublisher {
                             credentialProvider.updateCredential(mechanism, userName, change.get().toCredential());
                         else
                             credentialProvider.removeCredentials(mechanism, userName);
+                        }
                     });
                 });
             }
