+++ server-common/src/main/java/org/apache/kafka/server/common/CredentialProvider.java
